refactor:#41 병원 상세 위치 정보 함수 위치 변경 및 관련 코드 정리
