+++ src/utils/api/hospitalDetail.ts
@@ -1,3 +1,7 @@
+import { TABLE } from '@/constants/supabaseTables';
+import { Location } from '@/types/map';
+import { supabase } from '../supabase/supabase';
+
 /**
  * 병원 상세 정보 조회를 위한 기본 URL 생성 함수
  */
@@ -39,3 +43,31 @@ const hospitalDetail = async (id: string, options?: { isServer?: boolean }) => {
 };
 
 export default hospitalDetail;
+
+/**
+ * 지정한 id값의 병원 위치(위도, 경도)를 반환하는 함수
+ * @param id string : hpid
+ * @returns hospitalLocation - 지정한 id값의 병원 위치(위도, 경도) 정보
+ */
+export const getHospitalDetailLocation = async (id: string) => {
+  try {
+    let hospitalLocation: Location = { lat: 0, lng: 0 };
+
+    const { data, error } = await supabase
+      .from(TABLE.HOSPITALS)
+      .select('lat, lng')
+      .eq('id', id);
+
+    if (error) throw error;
+    if (!data || data.length === 0) return null;
+
+    const location = data[0];
+
+    hospitalLocation = { lat: location.lat, lng: location.lng };
+    console.log(hospitalLocation);
+
+    return hospitalLocation;
+  } catch (error) {
+    console.error('병원 상세 위치정보 불러오기 오류');
+  }
+};
