+++ src/utils/api/hospitals.ts
@@ -66,31 +66,3 @@ export const getAllHospitalData = async () => {
     console.error('병원 데이터 불러오기 오류', error);
   }
 };
-
-/**
- * 지정한 id값의 병원 위치(위도, 경도)를 반환하는 함수
- * @param id string : hpid
- * @returns hospitalLocation - 지정한 id값의 병원 위치(위도, 경도) 정보
- */
-export const getHospitalDetailLocation = async (id: string) => {
-  try {
-    let hospitalLocation: Location = { lat: 0, lng: 0 };
-
-    const { data, error } = await supabase
-      .from(TABLE.HOSPITALS)
-      .select('lat, lng')
-      .eq('id', id);
-
-    if (error) throw error;
-    if (!data || data.length === 0) return null;
-
-    const location = data[0];
-
-    hospitalLocation = { lat: location.lat, lng: location.lng };
-    console.log(hospitalLocation);
-
-    return hospitalLocation;
-  } catch (error) {
-    console.error('병원 상세 위치정보 불러오기 오류');
-  }
-};
